# openapi.yml
...
openapi: 3.1.0
info:
  title: CubeAPI
  description: E2B-compatible sandbox API server.
  license:
    name: ''
  version: 0.1.0
servers:
- url: /cubeapi/v1
  description: CubeAPI SDK/dashboard surface (source of truth generated via `cube-api --export-openapi`)
paths:
...
  /sandboxes:
    get:
      tags:
      - handlers::sandboxes
      operationId: list_sandboxes
      parameters:
      - name: metadata
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Sandbox list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListedSandbox'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - handlers::sandboxes
      operationId: create_sandbox
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewSandbox'
        required: true
      responses:
        '201':
          description: Sandbox created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sandbox'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Template or snapshot not found (e.g. deleted/tombstoned)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: A referenced resource is in a conflicting state and cannot be used to create a sandbox
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /sandboxes/{sandboxID}:
    get:
      tags:
      - handlers::sandboxes
      operationId: get_sandbox
      parameters:
      - name: sandboxID
        in: path
        description: Sandbox identifier
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Sandbox detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SandboxDetail'
        '404':
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
      - handlers::sandboxes
      operationId: kill_sandbox
      parameters:
      - name: sandboxID
        in: path
        description: Sandbox identifier
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Sandbox deleted
        '404':
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '408':
          description: The existing standard API request timeout expired before the synchronous delete completed
        '409':
          description: Paused sandbox cannot be admitted for internal resume because node capacity or resource metadata is unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: Sandbox is pausing, another lifecycle operation is in progress, the Cubelet RPC has too little remaining time, or its internal resume could not be completed
          headers:
            Retry-After:
              schema:
                type: integer
                format: int64
                minimum: 0
              description: Seconds a client should wait before retrying DELETE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /sandboxes/{sandboxID}/connect:
    post:
      tags:
      - handlers::sandboxes
      operationId: connect_sandbox
      parameters:
      - name: sandboxID
        in: path
        description: Sandbox identifier
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectSandbox'
        required: true
      responses:
        '200':
          description: Sandbox connection info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sandbox'
        '400':
          description: Invalid timeout value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Sandbox cannot be connected during a conflicting lifecycle transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /sandboxes/{sandboxID}/logs:
    get:
      tags:
      - handlers::sandboxes
      operationId: get_sandbox_logs
      parameters:
      - name: sandboxID
        in: path
        description: Sandbox identifier
        required: true
        schema:
          type: string
      - name: start
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Sandbox logs (legacy shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SandboxLogs'
        '404':
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /sandboxes/{sandboxID}/network:
    put:
      tags:
      - handlers::sandboxes
      operationId: update_sandbox_network
      parameters:
      - name: sandboxID
        in: path
        description: Sandbox identifier
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSandboxNetworkRequest'
        required: true
      responses:
        '204':
          description: Network policy updated
        '400':
          description: Invalid network policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Sandbox is not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /sandboxes/{sandboxID}/pause:
    post:
      tags:
      - handlers::sandboxes
      operationId: pause_sandbox
      parameters:
      - name: sandboxID
        in: path
        description: Sandbox identifier
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Sandbox paused
        '404':
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Sandbox cannot be paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /sandboxes/{sandboxID}/refreshes:
    post:
      tags:
      - handlers::sandboxes
      operationId: refresh_sandbox
      parameters:
      - name: sandboxID
        in: path
        description: Sandbox identifier
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '204':
          description: Sandbox refreshed
        '400':
          description: Invalid duration value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /sandboxes/{sandboxID}/resume:
    post:
      tags:
      - handlers::sandboxes
      operationId: resume_sandbox
      parameters:
      - name: sandboxID
        in: path
        description: Sandbox identifier
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumedSandbox'
        required: true
      responses:
        '201':
          description: Sandbox resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sandbox'
        '400':
          description: Invalid timeout value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Sandbox is already running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /templates/compat:
    get:
      tags:
      - handlers::templates
      operationId: template_compat
      responses:
        '200':
          description: Template compatibility matrix
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateCompatMatrixView'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    ConnectSandbox:
      type: object
      description: Request body for POST /sandboxes/{id}/connect.
      properties:
        timeout:
          type:
          - integer
          - 'null'
          format: int32
          description: Idle timeout in seconds; omitted to keep the current value, -1 for no expiry, and a positive value ensures at least that much remaining lifetime. For an already-running sandbox, a shorter positive window is ignored; use POST /sandboxes/{id}/timeout to shorten it. Zero and values below -1 are invalid.
...
    Sandbox:
      type: object
      description: |-
        Response for POST /sandboxes and POST /sandboxes/{id}/connect.
        All ID abbreviations uppercase per E2B OpenAPI spec.
      required:
      - templateID
      - sandboxID
      - clientID
      - envdVersion
      properties:
        alias:
          type:
          - string
          - 'null'
        clientID:
          type: string
        domain:
          type:
          - string
          - 'null'
        envdAccessToken:
          type:
          - string
          - 'null'
        envdVersion:
          type: string
        sandboxID:
          type: string
        templateID:
          type: string
        trafficAccessToken:
          type:
          - string
          - 'null'
...
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...